requirements.txt
numpy
matplotlib
scipy
Pillow

